Renames folder (editor to core)
