+++ InitGui.py
@@ -40,7 +40,7 @@ class NodesWorkbench(Workbench):
 
     def Initialize(self):
         # FC needs lazy import at this point
-        from editor.nodes_window import FCNWindow
+        from core.nodes_window import FCNWindow
 
         # Prepare the window
         self.window = FCNWindow()
