+++ core/nodes_default_node.py
