+++ core/nodes_drag_listbox.py
@@ -4,7 +4,7 @@ from qtpy.QtWidgets import QListWidget, QAbstractItemView, QListWidgetItem
 
 from nodeeditor.utils import dumpException
 
-from editor.nodes_conf import NodesStore, LISTBOX_MIMETYPE
+from core.nodes_conf import NodesStore, LISTBOX_MIMETYPE
 
 
 class QDMDragListbox(QListWidget):
