+++ nodes/generators/generator_solid_box.py
@@ -26,10 +26,10 @@ import FreeCAD as App
 import Part
 import awkward as ak
 
-from editor.nodes_conf import register_node
-from editor.nodes_base_node import FCNNode
+from core.nodes_conf import register_node
+from core.nodes_base_node import FCNNode
 from nodes_locator import icon
-from editor.nodes_utils import simplify, map_objects
+from core.nodes_utils import simplify, map_objects
 
 
 @register_node
