+++ nodes/list/list_next.py
@@ -22,8 +22,8 @@
 #
 #
 ###################################################################################
-from editor.nodes_conf import register_node
-from editor.nodes_base_node import FCNNode
+from core.nodes_conf import register_node
+from core.nodes_base_node import FCNNode
 
 import nodes_locator as locator
