+++ nodes/number/number_number_range.py
@@ -25,8 +25,8 @@
 import awkward as ak
 import numpy as np
 
-from editor.nodes_conf import register_node
-from editor.nodes_base_node import FCNNode
+from core.nodes_conf import register_node
+from core.nodes_base_node import FCNNode
 from nodes_locator import icon
