+++ nodes/number/number_scalar_math.py
@@ -24,8 +24,8 @@
 ###################################################################################
 import awkward as ak
 
-from editor.nodes_conf import register_node
-from editor.nodes_base_node import FCNNode
+from core.nodes_conf import register_node
+from core.nodes_base_node import FCNNode
 from nodes_locator import icon
