+++ nodes/scene/scene_timer.py
@@ -24,8 +24,8 @@
 ###################################################################################
 from qtpy import QtCore
 
-from editor.nodes_conf import register_node
-from editor.nodes_base_node import FCNNode
+from core.nodes_conf import register_node
+from core.nodes_base_node import FCNNode
 import nodes_locator as locator
